requirements.txt
selenium>=4.12.0
webdriver-manager>=3.8.6
python-dotenv>=1.0.0
mtranslate>=1.8
Pillow>=9.0.0
Flask
flask-cors
tensorflow
keras
opencv-python
speechrecognition==3.10.0
pyaudio==0.2.11
googletrans==4.0.0rc1
textblob==0.17.1
language-tool-python==2.7.1
pydub==0.25.1
requests==2.31.0
wave==0.0.2
nltk==3.8.1
pydub
google-cloud-translate